+++ vertigo-rest/src/main/java/io/vertigo/rest/RestfulService.java
@@ -1,0 +1,5 @@
+package io.vertigo.rest;
+
+public interface RestfulService {
+
+}